Python Project 3( Third Empire Decision)/requirements.txt
imutils==0.5.3
numpy==1.19.1
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
